Fix repo build for M1 MacBooks
