+++ package.json
@@ -38,14 +38,14 @@
     "@babel/cli": "7.20.7",
     "@babel/preset-env": "7.20.2",
     "@types/jest": "29.4.0",
-    "canvas": "2.11.0",
     "eslint": "8.33.0",
     "eslint-config-airbnb": "19.0.4",
     "eslint-plugin-import": "2.27.5",
     "eslint-plugin-jest": "27.2.1",
     "eslint-plugin-jsx-a11y": "6.7.1",
     "husky": "8.0.3",
-    "jest": "29.4.1"
+    "jest": "29.4.1",
+    "pngjs": "^7.0.0"
   },
   "engines": {
     "node": ">=16.15.0",
